pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dashscope>=1.25.17",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-openai>=0.3.0",
    "openai>=1.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pypdf>=6.10.2",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.14.0",
    "sse-starlette>=2.0.0",
    "streamlit>=1.57.0",
    "minio>=7.2.0",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
